seeds avoiding mongo bug
There is a bug in MongoDB which is fully hit with Mongoid when
data structure contains deeply nested arrays, like we have
Events > Competitions > Entries > Runs - three level deep
nested arrays inside an event.

The bug I found described here:
http://stackoverflow.com/questions/25609262 and the
request in MongoDB Jira is here:
jira.mongodb.org/browse/SERVER-831 (they consider it a
feature missing, but it is clearly a bug in design of
MongoDB and in implementation of Mongoid)
and is best understood on this example:
gist.github.com/durran/342738.

They wai it is solved is that happily even big Mongoid
objects can be built in memoty (using build instead of
create methods) which will then be saved in the end.

So what we do is: create Events straight into the DB, but
populate competitions, entries, races and runs in memory
and only when all are populated, we save them all.

Signed-off-by: Zeljko <zeljko@zwr.fi>